tentativa de usar o autorizacao com facades falhou

+++ app/Http/Controllers/CategoriaController.php
@@ -8,19 +8,20 @@ use Illuminate\Support\Facades\Storage;
 
 class CategoriaController extends Controller {
     public function index() {
-        // $this->authorize('index', Jogo::class);
+        $this->authorize('index', Categoria::class);
         $data = Categoria::all();    
-        return view('categoria.index', compact(['data']));
+        return view('categoria.index', compact('data'));
     }
+    
 
     public function create() {
-        // $this->authorize('create', Jogo::class);
+        $this->authorize('index', Categoria::class);
         $data = Categoria::all();
         return view('categoria.create', compact('data'));
     }
 
     public function store(Request $request) {
-        // $this->authorize('create', Jogo::class);
+        $this->authorize('index', Categoria::class);
         $data = new Categoria();
         $data->name = $request->nome;
         $data->descricao = $request->descricao;
@@ -29,7 +30,7 @@ class CategoriaController extends Controller {
     }
 
     public function show($id) {
-        // $this->authorize('show', Jogo::class);
+        $this->authorize('index', Categoria::class);
         $data = Categoria::find($id);
         if(isset($data)) {
             return view('categoria.show', compact(['data']));
@@ -38,7 +39,7 @@ class CategoriaController extends Controller {
     }
 
     public function edit($id) {
-        // $this->authorize('edit', Jogo::class);
+        $this->authorize('index', Categoria::class);
         $data = Categoria::find($id);
 
         if(isset($data)) {
@@ -48,7 +49,7 @@ class CategoriaController extends Controller {
     }
 
     public function update(Request $request, $id) {
-        // $this->authorize('edit', Jogo::class);
+        $this->authorize('index', Categoria::class);
         $data = Categoria::find($id);
         if(isset($data)) {
             $data->name = $request->nome;
@@ -60,7 +61,7 @@ class CategoriaController extends Controller {
     }
 
     public function destroy($id) {
-        // $this->authorize('destroy', Jogo::class);
+        $this->authorize('index', Categoria::class);
         $data = Categoria::find($id);
         if(isset($data)) {
             $data->delete();
